better handle quoted columns and tables
Tables and columns were not properly quoted in all instances when the
quote_char constructor option is set.  Some were underquoted, some over quooted,
and some table expressions were quoted.  This lead iin the latter cases to illegal SQL.

The intrinsic problem lies with treating them as strings, rather than
as objects, so there is no "quote" or "quotable" state assigned to a
given SQL fragment.  The code has to rely upon heuristics to know when
to quote (e.g. a bare identifier) and when not to quote (interpolation
of an SQL expression).  This requires carrying external state (e.g.,
the new is_literal flag from _parse_table and in the return from
table_alias), and informing the quoting apparati in table_alias() and
column_alias() of whether quoting is permissible via the "quote_name"
and "quote_aliased_name" options.